requirements.txt
ai2-tango==1.2.1
aiohttp==3.8.5
aiosignal==1.3.1
allennlp-light @ git+https://github.com/tchayintr/allennlp-light.git@1f6dc7cf9c0ae01b42d53b1b022a3dcd42147466
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
base58==2.1.1
blessed==1.20.0
boto3==1.28.43
botocore==1.31.43
cached-path==1.4.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-help-colors==0.9.2
cmake==3.25.0
decorator==5.1.1
dill==0.3.7
exceptiongroup==1.1.3
executing==1.2.0
fasttext-wheel==0.9.2
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.9.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.35
glob2==0.7
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
gpustat==1.1.1
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
lightning-lite==1.8.6
lightning-utilities==0.9.0
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.108
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
parso==0.8.3
petname==2.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==2.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pyg-lib==0.2.0+pt20cu118
Pygments==2.16.1
pynvim==0.4.3
pyparsing==3.1.1
pythainlp==4.0.2
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rjsonnet==0.5.3
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
smmap==5.0.0
sqlitedict==2.1.0
stack-data==0.6.2
sympy==1.12
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.12.1
torch==2.0.1+cu118
torch-cluster==1.6.1+pt20cu118
torch-geometric==2.3.1
torch-scatter==2.1.1+pt20cu118
torch-sparse==0.6.17+pt20cu118
torch-spline-conv==1.2.2+pt20cu118
torchaudio==2.0.2+cu118
torchmetrics==1.1.1
torchvision==0.15.2+cu118
tqdm==4.66.1
traitlets==5.9.0
transformers==4.20.1
triton==2.0.0
typing_extensions==4.7.1
urllib3==1.26.16
wcwidth==0.2.6
xxhash==3.3.0
yarl==1.9.2
